+++ odbc-api/src/buffers.rs
@@ -8,7 +8,7 @@ use odbc_sys::{CDataType, Len, Pointer};
 pub use self::{
     fixed_sized::{
         FixedSizedCType, OptDateColumn, OptF32Column, OptF64Column, OptFixedSizedColumn,
-        OptI32Column, OptI64Column, OptTimeColumn, OptTimestampColumn,
+        OptI32Column, OptI64Column, OptNumericColumn, OptTimeColumn, OptTimestampColumn,
     },
     text_column::TextColumn,
     text_row_set::TextRowSet,